requirements.txt
geopandas==1.0.1
pandas==2.2.3
rioxarray==0.17.0
xarray==2024.7.0
azure-storage-blob==12.22.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pyarrow==19.0.0
dask>2024.8.0
tqdm==4.66.1
matplotlib==3.8.2
ocha-stratus==0.1.2
dash-bootstrap-components==1.6.0
dash-extensions==1.0.18
dash-leaflet==1.0.15
dash-mantine-components==0.12.1
